requirements.txt
flask
Flask-SQLAlchemy
PyMySQL
requests
python-dotenv
cryptography
aiohttp
flask-login
flask-bcrypt
flask-limiter
# --upgrade google-api-python-client google-auth-httplib2 google-auth-oauthlib

# csv
rapidfuzz
# re